requirements.txt
torch>=2.0.0
transformers>=4.38.0
bitsandbytes>=0.43.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.28.0
pydantic>=2.5.0
pyyaml>=6.0.0
pytest>=7.4.0